Fix initial loading state with useMutation and add optmization (#39)
This makes sure that `loading` is set to false when `useMutation` is called. It causes issues having `loading` default to true for a mutation because it doesn't actually send the request until the consumer calls the returned method. 

This is probably considered a breaking change? @bmullan91 / @jackdclark 

Another small thing in this PR is an additional check to avoid making a new state object if we don't need to, this means react won't try to render wastefully.