Fixed an error which would happen if variables was undefined
I found that in an example with useMutation graphql-hooks would
call cache.get but the variables property on the key object was
undefined, this was causing graphql-norm to throw an error when
attempting to denormalize. This commit fixes this by passing in an
empty object instead if the variables property is undefined.