Use `any` for RefetchAfterMutationItem filter param (#986)
With `object` as the type, it's cumbersome to type it correctly (can't pass a type parameter, can't just index on it, have to use a typecast in the function body). This commit makes it so we just use any as the variables type, as that is actually more what we want. From our perspective, we pass `filter` a parameter and want a boolean back.